backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
transformers==4.35.0
opencv-python-headless==4.8.1.78
pillow==10.0.1
onnxruntime==1.16.3
numpy==1.24.3
gunicorn==21.2.0
